ai-inference/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
langchain==0.0.335
langchain-community==0.0.13
langchain-openai==0.0.5
openai==1.3.5
tiktoken==0.5.1
faiss-cpu==1.7.4
markdown==3.5.1
beautifulsoup4==4.12.2
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1 